refactor: 팔로우/언팔로우 예외 수정

+++ src/main/java/com/leets/team2/xclone/domain/follow/service/FollowService.java
@@ -38,7 +38,10 @@ public class FollowService {
     public void followUser(FollowDTO.Save dto, String myTag) {
         Member followee = memberRepository.findByTag(dto.tag()).orElseThrow(NoSuchMemberException::new);
         Member follower = memberRepository.findByTag(myTag).orElseThrow(NoSuchMemberException::new);
-
+        // 자기 자신을 팔로우하려고 하는 경우를 바로 예외 처리
+        if (follower.equals(followee)) {
+            throw new InvalidFollowException();
+        }
         Follow follows = followRepository.findByFollowee_TagAndFollower_Tag(dto.tag(), myTag);
         checkValidFollow(follows, 0);
 
@@ -54,6 +57,10 @@ public class FollowService {
         Member followee = memberRepository.findByTag(dto.tag()).orElseThrow(NoSuchMemberException::new);
         Member follower = memberRepository.findByTag(myTag).orElseThrow(NoSuchMemberException::new);
 
+        // 자기 자신을 언팔로우하려고 하는 경우를 바로 예외 처리
+        if (follower.equals(followee)) {
+            throw new InvalidFollowException();
+        }
         Follow followInfo = followRepository.findByFollowee_TagAndFollower_Tag(dto.tag(), myTag);
         checkValidFollow(followInfo, 1);
         followRepository.delete(followInfo);
@@ -63,10 +70,7 @@ public class FollowService {
     // 0 : follow
     // 1 : unfollow
     private void checkValidFollow(Follow followInfo, int option){
-        if(followInfo.getFollower().equals(followInfo.getFollowee())){
-            throw new InvalidFollowException(); // 자기 자신을 팔로우/언팔로우한 경우
-        }
-        if (!(followInfo ==null) &option==0) {
+        if (!(followInfo ==null) && option==0) {
             throw new FollowAlreadyExistsException();   // 이미 팔로우한 경우
         }
         if(followInfo==null&&option==1){
